+++ CONTRIBUTING.md
@@ -93,7 +93,7 @@ grab and store an openapi yaml spec from a target slurm version.
 
 I do this for every new SLURM version, so it should already be done.
 
-Assuming 24.05:
+Assuming 23.11:
 
 ```bash
 openapi-generator-cli generate \
@@ -106,6 +106,15 @@ openapi-generator-cli generate \
 ```
 
 This will generate an entire git repository that you can toss up in GitHub.
+
+Navigate to it.
+
+Now we just strip out all instances of the API version. In this case, 23.11 is API
+version 0.0.40:
+
+```bash
+find . -type f -not -path "./.git/*" -exec gsed -i 's/V0040//g' {} +
+```
 
 ### Cutting releases
