+++ internal/slurm/node_2311.go
@@ -1,0 +1,180 @@
+//go:build 2311
+
+package slurm
+
+import (
+	"context"
+	"fmt"
+	"log/slog"
+
+	"github.com/akyoto/cache"
+	openapi "github.com/lcrownover/openapi-slurm-23-11"
+	"github.com/lcrownover/prometheus-slurm-exporter/internal/api"
+	"github.com/lcrownover/prometheus-slurm-exporter/internal/types"
+	"github.com/prometheus/client_golang/prometheus"
+)
+
+type NodeCollector struct {
+	ctx      context.Context
+	cpuAlloc *prometheus.Desc
+	cpuIdle  *prometheus.Desc
+	cpuOther *prometheus.Desc
+	cpuTotal *prometheus.Desc
+	memAlloc *prometheus.Desc
+	memTotal *prometheus.Desc
+}
+
+// NewNodeCollectorOld creates a Prometheus collector to keep all our stats in
+// It returns a set of collections for consumption
+func NewNodeCollector(ctx context.Context) *NodeCollector {
+	labels := []string{"node", "status"}
+
+	return &NodeCollector{
+		ctx:      ctx,
+		cpuAlloc: prometheus.NewDesc("slurm_node_cpu_alloc", "Allocated CPUs per node", labels, nil),
+		cpuIdle:  prometheus.NewDesc("slurm_node_cpu_idle", "Idle CPUs per node", labels, nil),
+		cpuOther: prometheus.NewDesc("slurm_node_cpu_other", "Other CPUs per node", labels, nil),
+		cpuTotal: prometheus.NewDesc("slurm_node_cpu_total", "Total CPUs per node", labels, nil),
+		memAlloc: prometheus.NewDesc("slurm_node_mem_alloc", "Allocated memory per node", labels, nil),
+		memTotal: prometheus.NewDesc("slurm_node_mem_total", "Total memory per node", labels, nil),
+	}
+}
+
+// Send all metric descriptions
+func (nc *NodeCollector) Describe(ch chan<- *prometheus.Desc) {
+	ch <- nc.cpuAlloc
+	ch <- nc.cpuIdle
+	ch <- nc.cpuOther
+	ch <- nc.cpuTotal
+	ch <- nc.memAlloc
+	ch <- nc.memTotal
+}
+
+func (nc *NodeCollector) Collect(ch chan<- prometheus.Metric) {
+	apiCache := nc.ctx.Value(types.ApiCacheKey).(*cache.Cache)
+	nodeRespBytes, found := apiCache.Get("nodes")
+	if !found {
+		slog.Error("failed to get nodes response for cpu metrics from cache")
+		return
+	}
+	nodesResp, err := api.UnmarshalNodesResponse(nodeRespBytes.([]byte))
+	if err != nil {
+		slog.Error("failed to unmarshal nodes response for cpu metrics", "error", err)
+		return
+	}
+	nm, err := ParseNodeMetrics(*nodesResp)
+	if err != nil {
+		slog.Error("failed to collect nodes metrics", "error", err)
+		return
+	}
+	for node := range nm {
+		ch <- prometheus.MustNewConstMetric(nc.cpuAlloc, prometheus.GaugeValue, float64(nm[node].cpuAlloc), node, nm[node].nodeStatus)
+		ch <- prometheus.MustNewConstMetric(nc.cpuIdle, prometheus.GaugeValue, float64(nm[node].cpuIdle), node, nm[node].nodeStatus)
+		ch <- prometheus.MustNewConstMetric(nc.cpuOther, prometheus.GaugeValue, float64(nm[node].cpuOther), node, nm[node].nodeStatus)
+		ch <- prometheus.MustNewConstMetric(nc.cpuTotal, prometheus.GaugeValue, float64(nm[node].cpuTotal), node, nm[node].nodeStatus)
+		ch <- prometheus.MustNewConstMetric(nc.memAlloc, prometheus.GaugeValue, float64(nm[node].memAlloc), node, nm[node].nodeStatus)
+		ch <- prometheus.MustNewConstMetric(nc.memTotal, prometheus.GaugeValue, float64(nm[node].memTotal), node, nm[node].nodeStatus)
+	}
+}
+
+// NodeMetrics stores metrics for each node
+type nodeMetrics struct {
+	memAlloc   uint64
+	memTotal   uint64
+	cpuAlloc   uint64
+	cpuIdle    uint64
+	cpuOther   uint64
+	cpuTotal   uint64
+	nodeStatus string
+}
+
+func NewNodeMetrics() *nodeMetrics {
+	return &nodeMetrics{}
+}
+
+// ParseNodeMetrics takes the output of sinfo with node data
+// It returns a map of metrics per node
+func ParseNodeMetrics(nodesResp openapi.V0040OpenapiNodesResp) (map[string]*nodeMetrics, error) {
+	nodeMap := make(map[string]*nodeMetrics)
+
+	for _, n := range nodesResp.Nodes {
+		nodeName := *n.Hostname
+		nodeMap[nodeName] = &nodeMetrics{0, 0, 0, 0, 0, 0, ""}
+
+		// state
+		nodeStatesStr, err := GetNodeStatesString(n, "|")
+		if err != nil {
+			return nil, fmt.Errorf("failed to get node state: %v", err)
+		}
+		nodeMap[nodeName].nodeStatus = nodeStatesStr
+
+		// memory
+		nodeMap[nodeName].memAlloc = GetNodeAllocMemory(n)
+		nodeMap[nodeName].memTotal = GetNodeTotalMemory(n)
+
+		// cpu
+		nodeMap[nodeName].cpuAlloc = GetNodeAllocCPUs(n)
+		nodeMap[nodeName].cpuIdle = GetNodeIdleCPUs(n)
+		nodeMap[nodeName].cpuOther = GetNodeOtherCPUs(n)
+		nodeMap[nodeName].cpuTotal = GetNodeTotalCPUs(n)
+	}
+
+	return nodeMap, nil
+}
+
+type nodesMetrics struct {
+	alloc float64
+	comp  float64
+	down  float64
+	drain float64
+	err   float64
+	fail  float64
+	idle  float64
+	maint float64
+	mix   float64
+	resv  float64
+}
+
+func NewNodesMetrics() *nodesMetrics {
+	return &nodesMetrics{}
+}
+
+// ParseNodesMetrics iterates through node response objects and tallies up
+// nodes based on their state
+func ParseNodesMetrics(nodesResp openapi.V0040OpenapiNodesResp) (*nodesMetrics, error) {
+	nm := NewNodesMetrics()
+
+	for _, n := range nodesResp.Nodes {
+		nodeStates, err := GetNodeStates(n)
+		if err != nil {
+			return nil, fmt.Errorf("failed to get node state for nodes metrics: %v", err)
+		}
+
+		for _, ns := range *nodeStates {
+			switch ns {
+			case types.NodeStateAlloc:
+				nm.alloc += 1
+			case types.NodeStateComp:
+				nm.comp += 1
+			case types.NodeStateDown:
+				nm.down += 1
+			case types.NodeStateDrain:
+				nm.drain += 1
+			case types.NodeStateErr:
+				nm.err += 1
+			case types.NodeStateFail:
+				nm.fail += 1
+			case types.NodeStateIdle:
+				nm.idle += 1
+			case types.NodeStateMaint:
+				nm.maint += 1
+			case types.NodeStateMix:
+				nm.mix += 1
+			case types.NodeStateResv:
+				nm.resv += 1
+			}
+		}
+	}
+
+	return nm, nil
+}
