+++ internal/slurm/scheduler_2405.go
@@ -1,0 +1,185 @@
+//go:build 2405
+
+package slurm
+
+import (
+	"context"
+	"log/slog"
+
+	"github.com/akyoto/cache"
+	openapi "github.com/lcrownover/openapi-slurm-24-05"
+	"github.com/lcrownover/prometheus-slurm-exporter/internal/api"
+	"github.com/lcrownover/prometheus-slurm-exporter/internal/types"
+	"github.com/lcrownover/prometheus-slurm-exporter/internal/util"
+	"github.com/prometheus/client_golang/prometheus"
+)
+
+type SchedulerCollector struct {
+	ctx                               context.Context
+	threads                           *prometheus.Desc
+	queue_size                        *prometheus.Desc
+	dbd_queue_size                    *prometheus.Desc
+	last_cycle                        *prometheus.Desc
+	mean_cycle                        *prometheus.Desc
+	cycle_per_minute                  *prometheus.Desc
+	backfill_last_cycle               *prometheus.Desc
+	backfill_mean_cycle               *prometheus.Desc
+	backfill_depth_mean               *prometheus.Desc
+	total_backfilled_jobs_since_start *prometheus.Desc
+	total_backfilled_jobs_since_cycle *prometheus.Desc
+	total_backfilled_heterogeneous    *prometheus.Desc
+}
+
+func NewSchedulerCollector(ctx context.Context) *SchedulerCollector {
+	return &SchedulerCollector{
+		ctx: ctx,
+		threads: prometheus.NewDesc(
+			"slurm_scheduler_threads",
+			"Information provided by the Slurm sdiag command, number of scheduler threads ",
+			nil,
+			nil),
+		queue_size: prometheus.NewDesc(
+			"slurm_scheduler_queue_size",
+			"Information provided by the Slurm sdiag command, length of the scheduler queue",
+			nil,
+			nil),
+		dbd_queue_size: prometheus.NewDesc(
+			"slurm_scheduler_dbd_queue_size",
+			"Information provided by the Slurm sdiag command, length of the DBD agent queue",
+			nil,
+			nil),
+		last_cycle: prometheus.NewDesc(
+			"slurm_scheduler_last_cycle",
+			"Information provided by the Slurm sdiag command, scheduler last cycle time in (microseconds)",
+			nil,
+			nil),
+		mean_cycle: prometheus.NewDesc(
+			"slurm_scheduler_mean_cycle",
+			"Information provided by the Slurm sdiag command, scheduler mean cycle time in (microseconds)",
+			nil,
+			nil),
+		cycle_per_minute: prometheus.NewDesc(
+			"slurm_scheduler_cycle_per_minute",
+			"Information provided by the Slurm sdiag command, number scheduler cycles per minute",
+			nil,
+			nil),
+		backfill_last_cycle: prometheus.NewDesc(
+			"slurm_scheduler_backfill_last_cycle",
+			"Information provided by the Slurm sdiag command, scheduler backfill last cycle time in (microseconds)",
+			nil,
+			nil),
+		backfill_mean_cycle: prometheus.NewDesc(
+			"slurm_scheduler_backfill_mean_cycle",
+			"Information provided by the Slurm sdiag command, scheduler backfill mean cycle time in (microseconds)",
+			nil,
+			nil),
+		backfill_depth_mean: prometheus.NewDesc(
+			"slurm_scheduler_backfill_depth_mean",
+			"Information provided by the Slurm sdiag command, scheduler backfill mean depth",
+			nil,
+			nil),
+		total_backfilled_jobs_since_start: prometheus.NewDesc(
+			"slurm_scheduler_backfilled_jobs_since_start_total",
+			"Information provided by the Slurm sdiag command, number of jobs started thanks to backfilling since last slurm start",
+			nil,
+			nil),
+		total_backfilled_jobs_since_cycle: prometheus.NewDesc(
+			"slurm_scheduler_backfilled_jobs_since_cycle_total",
+			"Information provided by the Slurm sdiag command, number of jobs started thanks to backfilling since last time stats where reset",
+			nil,
+			nil),
+		total_backfilled_heterogeneous: prometheus.NewDesc(
+			"slurm_scheduler_backfilled_heterogeneous_total",
+			"Information provided by the Slurm sdiag command, number of heterogeneous job components started thanks to backfilling since last Slurm start",
+			nil,
+			nil),
+	}
+}
+
+// Send all metric descriptions
+func (c *SchedulerCollector) Describe(ch chan<- *prometheus.Desc) {
+	ch <- c.threads
+	ch <- c.queue_size
+	ch <- c.dbd_queue_size
+	ch <- c.last_cycle
+	ch <- c.mean_cycle
+	ch <- c.cycle_per_minute
+	ch <- c.backfill_last_cycle
+	ch <- c.backfill_mean_cycle
+	ch <- c.backfill_depth_mean
+	ch <- c.total_backfilled_jobs_since_start
+	ch <- c.total_backfilled_jobs_since_cycle
+	ch <- c.total_backfilled_heterogeneous
+}
+
+// Send the values of all metrics
+func (sc *SchedulerCollector) Collect(ch chan<- prometheus.Metric) {
+	apiCache := sc.ctx.Value(types.ApiCacheKey).(*cache.Cache)
+	diagRespBytes, found := apiCache.Get("diag")
+	if !found {
+		slog.Error("failed to get diag response for scheduler metrics from cache")
+		return
+	}
+	diagResp, err := api.UnmarshalDiagResponse(diagRespBytes.([]byte))
+	if err != nil {
+		slog.Error("failed to unmarshal diag response for scheduler metrics", "error", err)
+		return
+	}
+	sm, err := ParseSchedulerMetrics(*diagResp)
+	if err != nil {
+		slog.Error("failed to collect scheduler metrics", "error", err)
+		return
+	}
+	ch <- prometheus.MustNewConstMetric(sc.threads, prometheus.GaugeValue, sm.threads)
+	ch <- prometheus.MustNewConstMetric(sc.queue_size, prometheus.GaugeValue, sm.queue_size)
+	ch <- prometheus.MustNewConstMetric(sc.dbd_queue_size, prometheus.GaugeValue, sm.dbd_queue_size)
+	ch <- prometheus.MustNewConstMetric(sc.last_cycle, prometheus.GaugeValue, sm.last_cycle)
+	ch <- prometheus.MustNewConstMetric(sc.mean_cycle, prometheus.GaugeValue, sm.mean_cycle)
+	ch <- prometheus.MustNewConstMetric(sc.cycle_per_minute, prometheus.GaugeValue, sm.cycle_per_minute)
+	ch <- prometheus.MustNewConstMetric(sc.backfill_last_cycle, prometheus.GaugeValue, sm.backfill_last_cycle)
+	ch <- prometheus.MustNewConstMetric(sc.backfill_mean_cycle, prometheus.GaugeValue, sm.backfill_mean_cycle)
+	ch <- prometheus.MustNewConstMetric(sc.backfill_depth_mean, prometheus.GaugeValue, sm.backfill_depth_mean)
+	ch <- prometheus.MustNewConstMetric(sc.total_backfilled_jobs_since_start, prometheus.GaugeValue, sm.total_backfilled_jobs_since_start)
+	ch <- prometheus.MustNewConstMetric(sc.total_backfilled_jobs_since_cycle, prometheus.GaugeValue, sm.total_backfilled_jobs_since_cycle)
+	ch <- prometheus.MustNewConstMetric(sc.total_backfilled_heterogeneous, prometheus.GaugeValue, sm.total_backfilled_heterogeneous)
+}
+
+func NewSchedulerMetrics() *schedulerMetrics {
+	return &schedulerMetrics{}
+}
+
+type schedulerMetrics struct {
+	threads                           float64
+	queue_size                        float64
+	dbd_queue_size                    float64
+	last_cycle                        float64
+	mean_cycle                        float64
+	cycle_per_minute                  float64
+	backfill_last_cycle               float64
+	backfill_mean_cycle               float64
+	backfill_depth_mean               float64
+	total_backfilled_jobs_since_start float64
+	total_backfilled_jobs_since_cycle float64
+	total_backfilled_heterogeneous    float64
+}
+
+// Extract the relevant metrics from the sdiag output
+func ParseSchedulerMetrics(diagResp openapi.SlurmV0041GetDiag200Response) (*schedulerMetrics, error) {
+	sm := NewSchedulerMetrics()
+	s := diagResp.Statistics
+
+	sm.threads = util.GetValueOrZero(s.ServerThreadCount)
+	sm.queue_size = util.GetValueOrZero(s.AgentQueueSize)
+	sm.dbd_queue_size = util.GetValueOrZero(s.DbdAgentQueueSize)
+	sm.last_cycle = util.GetValueOrZero(s.ScheduleCycleLast)
+	sm.mean_cycle = util.GetValueOrZero(s.ScheduleCycleMean)
+	sm.cycle_per_minute = util.GetValueOrZero(s.ScheduleCyclePerMinute)
+	sm.backfill_depth_mean = util.GetValueOrZero(s.BfDepthMean)
+	sm.backfill_last_cycle = util.GetValueOrZero(s.BfCycleLast)
+	sm.backfill_mean_cycle = util.GetValueOrZero(s.BfCycleMean)
+	sm.total_backfilled_jobs_since_cycle = util.GetValueOrZero(s.BfBackfilledJobs)
+	// TODO: This is probably not correct, should revisit this number
+	sm.total_backfilled_jobs_since_start = util.GetValueOrZero(s.BfLastBackfilledJobs)
+	sm.total_backfilled_heterogeneous = util.GetValueOrZero(s.BfBackfilledHetJobs)
+	return sm, nil
+}
